oneclerk/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "alembic==1.13.3",
    "asyncpg==0.30.0",
    "celery[redis]==5.4.0",
    "cryptography==43.0.3",
    "deepgram-sdk==3.7.0",
    "elevenlabs==1.9.0",
    "email-validator==2.2.0",
    "fastapi==0.115.0",
    "google-api-python-client==2.154.0",
    "google-auth==2.35.0",
    "google-auth-oauthlib==1.2.1",
    "gunicorn==22.0.0",
    "httpx==0.27.2",
    "openai==1.54.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "redis==5.2.0",
    "slowapi==0.1.9",
    "sqlalchemy[asyncio]==2.0.36",
    "stripe==11.1.0",
    "telnyx==2.1.1",
    "uvicorn[standard]==0.30.0",
    "websockets==13.1",
]
